requirements.txt
langchain
langchain-openai
python-dotenv
langchain_community
langchain_core
streamlit
openai
langchain-groq
chromadb
langchain-chroma
langchain-huggingface
arxiv
wikipedia
pathlib
streamlit
langchain_groq
sentence_transformers
pypdf
faiss-cpu
pymupdf